- Cleaned codes - Commented in some files - Made indents correct
