+++ iOS-Challenge/Coordinator/Coordinator.swift
@@ -8,9 +8,6 @@
 
 import UIKit
 
-
-import UIKit
-
 protocol Coordinator {
     var childCoordinators: [Coordinator] { get set }
     var navigationController: UINavigationController { get set }
