+++ iOS-Challenge/Model/RepoViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  RepoViewModel.swift
+//  iOS-Challenge
+//
+//  Created by iMamad on 6/5/20.
+//  Copyright © 2020 Farshad Mousalou. All rights reserved.
+//
+
+import Foundation
+
+struct RepoRowViewModel {
+    let nameText: String
+    let imageUrl: URL?
+    let starsCount: String
+    let commitsURL: String?
+}
