+++ iOS-Challenge/View/CommitsView.swift
@@ -8,7 +8,6 @@
 
 import UIKit
 
-
 class CommitsView: UIView {
     //MARK:- Init
     override init(frame: CGRect) {
