+++ iOS-Challenge/View/Search/SearchView.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import SDWebImage
 
 class SearchView: UIView {
     //MARK:- Init
@@ -61,36 +60,4 @@ class SearchView: UIView {
         barBtn.tintColor = .white
         return barBtn
     }()
-}
-
-
-extension UIImage {
-    enum ContentMode {
-        case contentFill
-        case contentAspectFill
-        case contentAspectFit
-    }
-
-    func resize(withSize size: CGSize, contentMode: ContentMode = .contentAspectFill) -> UIImage? {
-        let aspectWidth = size.width / self.size.width
-        let aspectHeight = size.height / self.size.height
-
-        switch contentMode {
-        case .contentFill:
-            return resize(withSize: size)
-        case .contentAspectFit:
-            let aspectRatio = min(aspectWidth, aspectHeight)
-            return resize(withSize: CGSize(width: self.size.width * aspectRatio, height: self.size.height * aspectRatio))
-        case .contentAspectFill:
-            let aspectRatio = max(aspectWidth, aspectHeight)
-            return resize(withSize: CGSize(width: self.size.width * aspectRatio, height: self.size.height * aspectRatio))
-        }
-    }
-
-    private func resize(withSize size: CGSize) -> UIImage? {
-        UIGraphicsBeginImageContextWithOptions(size, false, self.scale)
-        defer { UIGraphicsEndImageContext() }
-        draw(in: CGRect(x: 0.0, y: 0.0, width: size.width, height: size.height))
-        return UIGraphicsGetImageFromCurrentImageContext()
-    }
 }
